+++ src/test/data-contraction-all.ts
@@ -3,5 +3,6 @@ import {data as data_30} from './data-contraction-30';
 import {data as data_40} from './data-contraction-40';
 import {data as data_50} from './data-contraction-50';
 import {data as data_60} from './data-contraction-60';
+import {data as data_70} from './data-contraction-70';
 
-export const data = [...data_20, ...data_30, ...data_40, ...data_50, ...data_60];
+export const data = [...data_20, ...data_30, ...data_40, ...data_50, ...data_60, ...data_70];