+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - `#55`, as user, i want to optionally use all contractions.
 - `#56`, as user, i want to optionally use contraction 50.
 - `#57`, as user, i want to optionally use contraction 60.
+- `#58`, as user, i want to optionally use contraction 70.
 
 ## `1.8.0` - 5 July 2022